+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/AddVenusFacade.java
@@ -1,29 +1,22 @@
 package org.fugerit.java.doc.project.facade;
 
 import freemarker.cache.ClassTemplateLoader;
-import freemarker.cache.StringTemplateLoader;
 import freemarker.template.*;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.maven.model.Dependency;
-import org.apache.maven.model.DependencyManagement;
-import org.apache.maven.model.Model;
 import org.fugerit.java.core.cfg.ConfigException;
 import org.fugerit.java.core.function.SafeFunction;
 import org.fugerit.java.core.io.FileIO;
 import org.fugerit.java.core.io.StreamIO;
-import org.fugerit.java.core.javagen.JavaGenerator;
 import org.fugerit.java.core.javagen.SimpleJavaGenerator;
 import org.fugerit.java.core.lang.helpers.ClassHelper;
 import org.fugerit.java.doc.freemarker.config.FreeMarkerConfigStep;
-import org.fugerit.java.doc.freemarker.fun.SimpleMessageFun;
 import org.fugerit.java.doc.freemarker.process.FreemarkerDocProcessConfig;
 import org.fugerit.java.doc.freemarker.process.FreemarkerDocProcessConfigFacade;
-import org.maxxq.maven.dependency.ModelIO;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
+import java.sql.Timestamp;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Slf4j
 public class AddVenusFacade extends BasicVenusFacade {
@@ -34,6 +27,42 @@ public class AddVenusFacade extends BasicVenusFacade {
 
     private static final String LINE = "************************************************************************************************************************";
 
+    private static void processFile( String templatePath, File outputFile, Configuration configuration, Map<Object, Object> data ) throws IOException, TemplateException {
+        Template docTemplate = configuration.getTemplate( templatePath );
+        data.put( "templatePath", templatePath );
+        data.put( "generationTime", new Timestamp( System.currentTimeMillis() ) );
+        try ( Writer writer = new FileWriter( outputFile) ) {
+            docTemplate.process( data, writer );
+        }
+    }
+
+    private static File toFile( File base, String packageName, String fileName ) {
+        File folder = new File( base, packageName.replace( '.', '/' ) );
+        File file = new File( folder, fileName );
+        boolean mkdirs = Boolean.FALSE;
+        if ( !folder.exists() ) {
+            mkdirs = folder.mkdirs();
+        }
+        log.debug( "toFile : [{}], folder create : {} - {}", file.getAbsolutePath(), folder, mkdirs );
+        return file;
+    }
+
+    private static void addSampleStructure( VenusContext context, Configuration configuration, Map<Object, Object> data ) throws IOException, TemplateException {
+        if ( context.isAddJunit5() ) {
+            log.debug( "add junit5 structure" );
+            String fileBase = "DocHelperTest";
+            processFile( fileBase+".ftl",
+                    toFile( context.getTestJavaFolder(), "test."+context.getDocConfigPackage(), fileBase+".java" ),
+                    configuration, data );
+        } else {
+            log.debug( "create DocHelperExample class" );
+            String fileBase = "DocHelperExample";
+            processFile( fileBase+".ftl",
+                    toFile( context.getMainJavaFolder(), context.getDocConfigPackage(), fileBase+".java" ),
+                    configuration, data );
+        }
+    }
+
     private static void addDocFacade( VenusContext context ) throws IOException, TemplateException, ConfigException {
         // freemarker configuration
         Configuration configuration = new Configuration( new Version( FreeMarkerConfigStep.ATT_FREEMARKER_CONFIG_KEY_VERSION_LATEST ) );
@@ -43,17 +72,13 @@ public class AddVenusFacade extends BasicVenusFacade {
         configuration.setTemplateLoader( loader );
         configuration.setDefaultEncoding(StandardCharsets.UTF_8.name());
         // config generation
-        Template fmConfigTemplate = configuration.getTemplate( "fm-doc-process-config-template.ftl" );
         Map<Object, Object> data = new HashMap<>();
         data.put( "context" , context );
-        File fmConfigFile = new File( context.getProjectDir(), "src/main/resources/"+context.getResourcePathFmConfigXml() );
+        File fmConfigFile = new File( context.getMainResourcesFolder(), context.getResourcePathFmConfigXml() );
         log.info( "fmConfigFile : {}, mk parent? : {}", fmConfigFile.getCanonicalPath(), fmConfigFile.getParentFile().mkdirs() );
         File templateDir = new File( fmConfigFile.getParentFile(), context.getTemplateSubPath() );
         log.info( "templateDir : {}, mk parent? : {}", templateDir.getCanonicalPath(), templateDir.mkdirs() );
-        try ( Writer writer = new FileWriter( fmConfigFile ) ) {
-            fmConfigTemplate.process( data, writer );
-        }
-        configuration.clearTemplateCache();
+        processFile( "fm-doc-process-config-template.ftl", fmConfigFile, configuration, data );
         // copy sample template
         String fileName = "document.ftl";
         File documentExample = new File( templateDir, fileName );
@@ -62,7 +87,7 @@ public class AddVenusFacade extends BasicVenusFacade {
             FileIO.writeString( documentContent, documentExample );
         }
         // create doc config
-        File sourceFolder = new File( context.getProjectDir(), "src/main/java" );
+        File sourceFolder = context.getMainJavaFolder();
         log.info( "sourceFolder : {}, mk parent? : {}", sourceFolder.getCanonicalPath(), sourceFolder.mkdirs() );
         File resourceFolder = new File( context.getProjectDir(), "src/main/resources" );
         log.info( "resourceFolder : {}, mk parent? : {}", resourceFolder.getCanonicalPath(), resourceFolder.mkdirs() );
@@ -71,12 +96,8 @@ public class AddVenusFacade extends BasicVenusFacade {
         javaGenerator.init( sourceFolder, context.getDocConfigPackage()+"."+context.getDocConfigClass(), SimpleJavaGenerator.STYLE_CLASS, generatorProps);
         javaGenerator.generate();
         javaGenerator.write();
-        // create examples
-        Template docExampleTemplate = configuration.getTemplate( "DocHelperExample.ftl" );
-        File docExampleFile = new File( new File( sourceFolder, context.getDocConfigPackage().replace( '.', '/' ) ), "DocHelperExample.java" );
-        try ( Writer writer = new FileWriter( docExampleFile) ) {
-            docExampleTemplate.process( data, writer );
-        }
+        // add sample structure
+        addSampleStructure( context, configuration, data );
         if ( context.getModules().contains( "fj-doc-mod-fop" ) ) {
             String fopConfigName = "fop-config.xml";
             File fopConfig = new File( new File( resourceFolder, context.getArtificatIdForFolder() ),  fopConfigName );
@@ -85,6 +106,8 @@ public class AddVenusFacade extends BasicVenusFacade {
                 FileIO.writeString( fopConfigContent, fopConfig );
             }
         }
+        // clear template cache
+        configuration.clearTemplateCache();
     }
 
     public static boolean addVenusToMavenProject( VenusContext context ) {
@@ -95,7 +118,11 @@ public class AddVenusFacade extends BasicVenusFacade {
                 addExtensionList( pomFile, context );
                 if ( context.isAddDocFacace() ) {
                     addDocFacade( context );
-                    log.info( "Generation complete:\n{}\n* For usage open the example main() : {} *\n{}", LINE, context.getDocConfigPackage()+"."+context.getDocConfigClass()+"Example", LINE );
+                    if ( context.isAddJunit5() ) {
+                        log.info( "Generation complete:\n{}\n* For usage open the example junit : {} *\n{}", LINE, "test."+context.getDocConfigPackage()+"."+context.getDocConfigClass()+"Test", LINE );
+                    } else {
+                        log.info( "Generation complete:\n{}\n* For usage open the example main() : {} *\n{}", LINE, context.getDocConfigPackage()+"."+context.getDocConfigClass()+"Example", LINE );
+                    }
                     log.info( "for documentation refer to https://github.com/fugerit-org/fj-doc/blob/main/fj-doc-maven-plugin/README.md" );
                 }
             } else {